fix: fix tests
- Make a test for Epic.toString() to check all epic subtasks, not the
first one only.
- Move tests for an in-memory task manager instance from an abstract
super class to InMemoryTaskManagerTest.
- Add a test when a constructor of a file backed task manager receives
null instead of a HistoryManager instance.
- Make a constant WRONG_EXCEPTION_MESSAGE in an abstract task manager
superclass available for subclasses.
- Add a task end time check to assertTaskEquals().
- Rearrange constants in TestModels class.
